+++ test/real128_io.cpp
@@ -45,27 +45,17 @@ TEST_CASE("real128 io")
     std::uniform_int_distribution<int> sdist(0, 1);
     std::uniform_real_distribution<double> dist1(100., 1000.);
     for (int i = 0; i < ntries; ++i) {
-        check_round_trip(real128{::nextafterq(dist1(rng), 10000.) * (sdist(rng) ? 1 : -1)});
+        check_round_trip(nextafter(real128{dist1(rng)}, real128{10000.}) * (sdist(rng) ? 1 : -1));
     }
     std::uniform_real_distribution<double> dist2(1E-6, 1E-1);
     for (int i = 0; i < ntries; ++i) {
-        check_round_trip(real128{::nextafterq(dist2(rng), 1.) * (sdist(rng) ? 1 : -1)});
+        check_round_trip(nextafter(real128{dist2(rng)}, real128{1.}) * (sdist(rng) ? 1 : -1));
     }
     std::uniform_real_distribution<double> dist3(1E100, 1E120);
     for (int i = 0; i < ntries; ++i) {
-        check_round_trip(real128{::nextafterq(dist3(rng), 1E121) * (sdist(rng) ? 1 : -1)});
+        check_round_trip(nextafter(real128{dist3(rng)}, real128{1E121}) * (sdist(rng) ? 1 : -1));
     }
     // Some subnormals.
     check_round_trip(real128{"1E-4960"});
     check_round_trip(real128{"-1E-4960"});
-    // Small tests for the input stream.
-    real128 r;
-    std::istringstream iss;
-    iss.str("123.456");
-    iss >> r;
-    REQUIRE((r.m_value == real128{"123.456"}.m_value));
-    iss.clear();
-    iss.str("-1E-4000");
-    iss >> r;
-    REQUIRE((r.m_value == real128{"-1E-4000"}.m_value));
 }